// src/app/Types/approvals.ts
export class Approvals {
  Record: string;
  RcptDate: string;
  Source: string;
  AppUser: string;
  Status: string;
  Archivelink: string;
  InvoiceNumber: string;
  Company: string;
  Po: string;
  VendorNumber: string;
  VendorName: string;
  Curr: string;
  Amount: string;
  IvDocument: string;
  Document: string;
}